Zoom on real user location & Bugfix (wrong time shown in EventDetail)

+++ app/src/main/java/nl/imanidap/meet/MeetupEventsDownloadTask.java
@@ -45,7 +45,7 @@ public class MeetupEventsDownloadTask extends AsyncTask<String, Void, ArrayList<
                     meetEvent.setLatitude(event.getJSONObject("venue").getDouble("lat"));
                     meetEvent.setLongitude(event.getJSONObject("venue").getDouble("lon"));
                     meetEvent.setDescription(event.getString("description"));
-                    meetEvent.setTime(event.getInt("time"));
+                    meetEvent.setTime(event.getLong("time"));
                     meetEvent.setRsvpCount(event.getInt("yes_rsvp_count"));
 
                     meetEvents.add(meetEvent);
